+++ src/client/components/chatmenu/ChatMenu.spec.js
