+++ src/client/components/chatmenu/menu.scss
@@ -1,4 +1,4 @@
-@import '../../../mixins/_variables.scss';
+@import '../../mixins/_variables.scss';
 
 .chat-menu {
   position: relative;
